finishing writing

+++ include/fimex/CDMDimension.h
@@ -14,14 +14,14 @@ public:
 	CDMDimension(std::string name, long length);
 	virtual ~CDMDimension();
 	const std::string& getName() const {return name;}
-	long getLength() const {return length;}
+	size_t getLength() const {return length;}
 	void setUnlimited(int unlimited) {this->unlimited = unlimited;}
 	int isUnlimited() const {return unlimited;}
 	/// print xml representation to stream
 	void toXMLStream(std::ostream& out) const;
 private:
 	std::string name;
-	long length;
+	size_t length;
 	int unlimited;
 
 };
